refactor: 영속성 계층을 infrastructure로 이동
