+++ src/test/java/com/techfork/domain/recommendation/service/LlmRecommendationServiceTest.java
@@ -11,8 +11,8 @@ import com.techfork.global.elasticsearch.query.VectorQueryBuilder;
 import com.techfork.post.domain.projection.PostDocument;
 import com.techfork.post.fixture.PostDocumentFixture;
 import com.techfork.domain.recommendation.config.RecommendationProperties;
-import com.techfork.domain.recommendation.repository.RecommendedPostRepository;
-import com.techfork.domain.recommendation.repository.RecommendationHistoryRepository;
+import com.techfork.recommendation.infrastructure.RecommendedPostRepository;
+import com.techfork.recommendation.infrastructure.RecommendationHistoryRepository;
 import com.techfork.post.domain.Post;
 import com.techfork.global.util.TimeDecayStrategy;
 import com.techfork.personalization.application.query.lookup.PersonalizationProfileLookupItem;
