+++ src/test/java/com/techfork/recommendation/infrastructure/RecommendedPostRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.techfork.domain.recommendation.repository;
+package com.techfork.recommendation.infrastructure;
 
 import com.techfork.recommendation.domain.RecommendedPost;
 import com.techfork.domain.source.entity.TechBlog;
